分析从TCP返回的byte怎么解析到TraciResult，尝试将通信协定和数据类型提取出来减少相互依赖

+++ CodingConnected.TraCI.NET/Types/ClassType/TraCIBaseType.cs
@@ -7,8 +7,8 @@ namespace CodingConnected.TraCI.NET.Types;
 public class TraCIByte : ITraCIType
     {
     public byte TYPE { get; } = TraCIConstants.TYPE_BYTE;
-    public byte Value { get; set; }
-    public byte[] ToBytes() => [Value];
+    public SByte Value { get; set; }
+    public byte[] ToBytes() => [(byte)Value];
     }
 public class TraCIUByte : ITraCIType
     {
